+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # CHANGELOG
 
+## [UNRELEASED]
+
+### Added
+
+* Add paused param to start_playback command to allow to create player silently
+
+### Fixed
+
+* Fix issue checking parameters of add_tracks command
+
 ## [1.0.0] - 2021-11-26
 
 ### Added